more tweaks/refinements for superimposed groups

+++ gui/SGXSceneViewer/SceneViewer/glyphnode.h
@@ -69,7 +69,7 @@ namespace SynGlyphX
 
 		void setAlternatePosition( float group, const glm::vec3& pos ) const { alternate_position = pos; alternate_position_group = group; }
 		const glm::vec3& getAlternatePosition() const { return getRootParent()->alternate_position; }
-		const float getAlternatePositionGroup() const { return getRootParent()->alternate_position_group; }
+		const unsigned int getAlternatePositionGroup() const { return getRootParent()->alternate_position_group; }
 
 		const glm::mat4& getCachedTransform() const { return cached_transform; }
 		glm::vec3 getCachedPosition() const { return glm::vec3( cached_transform[3] ); }
@@ -122,7 +122,7 @@ namespace SynGlyphX
 		mutable render::sphere_bound bound, combined_bound;
 		mutable glm::vec3 visual_scale;
 		mutable glm::vec3 alternate_position;
-		mutable float alternate_position_group;
+		mutable unsigned int alternate_position_group;
 		mutable glm::vec3 animation_axis;
 		mutable float animation_rate;
